修正短信发送模板不起效的问题。 对企业微信通知添加invaliduser判断，为空才表示发送成功。

+++ Expand/weixinWork.php
@@ -55,7 +55,7 @@ class weixinWork {
         $result = json_decode($this->notice($param['send_account'], $param['send_content']), true);
 
         //发送成功，删除消息
-        if($result['errmsg'] == 'ok'){
+        if($result['errmsg'] == 'ok' && empty($result['invaliduser']) ){
             \Core\Func\CoreFunc::db('send')->where('send_id = :send_id')->delete([
                 'send_id' => $param['send_id']
             ]);
